Conflicts:
	coding/ows-v110/pom.xml
	coding/sos-v20/pom.xml
	core/cache/pom.xml
	core/pom.xml
	hibernate/h2/pom.xml
	misc/db/MySQL/series/MySQL_update_43_44_series_table.sql
	misc/db/Oracle/series/Oracle_437_to_438.sql
	misc/db/SqlServer/series/SqlServer_437_to_438.sql
	operations/core-v100/pom.xml
	operations/pom.xml
	pom.xml
	spring/views/src/main/webapp/WEB-INF/views/documentation.jsp

+++ extensions/do/core/src/main/java/org/n52/sos/ext/deleteobservation/DeleteObservationContentModificationListener.java
@@ -57,13 +57,22 @@ public class DeleteObservationContentModificationListener implements SosEventLis
     public void handle(SosEvent event) {
         if (event instanceof DeleteObservationEvent) {
             DeleteObservationEvent e = (DeleteObservationEvent) event;
-            DeleteObservationCacheControllerUpdate update =
-                    new DeleteObservationCacheControllerUpdate(e.getDeletedObservation());
-            LOGGER.debug("Updating Cache after content modification: {}", update);
-            try {
-                Configurator.getInstance().getCacheController().update(update);
-            } catch (OwsExceptionReport ex) {
-                LOGGER.error("Error processing Event", ex);
+            if (e.getDeletedObservation() != null) {
+                DeleteObservationCacheControllerUpdate update =
+                        new DeleteObservationCacheControllerUpdate(e.getDeletedObservation());
+                LOGGER.debug("Updating Cache after content modification: {}", update);
+                try {
+                    Configurator.getInstance().getCacheController().update(update);
+                } catch (OwsExceptionReport ex) {
+                    LOGGER.error("Error processing Event", ex);
+                }
+            } else {
+                LOGGER.debug("Updating Cache after content modification!");
+                try {
+                    Configurator.getInstance().getCacheController().update();
+                } catch (OwsExceptionReport ex) {
+                    LOGGER.error("Error processing Event", ex);
+                }
             }
         } else {
             LOGGER.debug("Can not handle modification event: {}", event);
